rename file: image snapshot configuration
